apps/book-picker/pyproject.toml
[project]
name = "book-picker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "requests>=2.32.3",
]
